+++ errhandler.go
@@ -1,0 +1,11 @@
+package main
+
+import "net/http"
+
+type errHandler func(http.ResponseWriter, *http.Request) *serverError
+
+func (fn errHandler) ServeHTTP(w http.ResponseWriter, r *http.Request) {
+	if err := fn(w, r); err != nil {
+		err.Log()
+	}
+}
