+++ main.go
@@ -1,11 +1,9 @@
 package main
 
 import (
-	"encoding/hex"
 	"encoding/json"
 	"github.com/Henry-Sarabia/blank"
 	"github.com/Henry-Sarabia/discovernow/views"
-	spotifyservice "github.com/Henry-Sarabia/refind/spotify"
 	"github.com/gorilla/handlers"
 	"github.com/gorilla/mux"
 	"github.com/gorilla/sessions"
@@ -20,13 +18,15 @@ import (
 )
 
 const (
-	APIURL          string = "http://127.0.0.1:3000/api/v1/"
-	redirectPath    string = "/results"
-	state           string = "abc123"
-	hashKeyName     string = "DISCOVER_HASH"
-	storeAuthName   string = "DISCOVER_AUTH"
-	storeCryptName  string = "DISCOVER_CRYPT"
-	frontendURIName string = "FRONTEND_URI"
+	apiPath          string = "/api/v1/"
+	loginEndpoint    string = "login"
+	playlistEndpoint string = "playlist"
+	redirectPath     string = "/results"
+	state            string = "abc123"
+	hashKeyName      string = "DISCOVER_HASH"
+	storeAuthName    string = "DISCOVER_AUTH"
+	storeCryptName   string = "DISCOVER_CRYPT"
+	frontendURIName  string = "FRONTEND_URI"
 )
 
 var (
@@ -41,58 +41,6 @@ var (
 	frontendURI string
 	auth        *spotify.Authenticator
 )
-
-func init() {
-	var err error
-
-	hashKey, err = decodeEnv(hashKeyName)
-	if err != nil {
-		log.Fatal(err)
-	}
-
-	storeAuth, err = decodeEnv(storeAuthName)
-	if err != nil {
-		log.Fatal(err)
-	}
-
-	storeCrypt, err = decodeEnv(storeCryptName)
-	if err != nil {
-		log.Fatal(err)
-	}
-
-	frontendURI, err = getEnv(frontendURIName)
-	if err != nil {
-		log.Fatal(err)
-	}
-
-	auth, err = spotifyservice.Authenticator(frontendURI + redirectPath)
-	if err != nil {
-		log.Fatalf("stack trace:\n%+v\n", err)
-	}
-}
-
-func decodeEnv(name string) (string, error) {
-	env, err := getEnv(name)
-	if err != nil {
-		return "", err
-	}
-
-	d, err := hex.DecodeString(env)
-	if err != nil {
-		return "", errors.Wrapf(err, "environment variable with name '%s' cannot be decoded from hex", name)
-	}
-
-	return string(d), nil
-}
-
-func getEnv(name string) (string, error) {
-	env, ok := os.LookupEnv(name)
-	if !ok {
-		return "", errors.Errorf("environment variable with name '%s' cannot be found", name)
-	}
-
-	return env, nil
-}
 
 func main() {
 	landing = views.NewView("index", "views/landing.gohtml")
@@ -121,26 +69,8 @@ func main() {
 	log.Fatal(srv.ListenAndServe())
 }
 
-type serverError struct {
-	Error   error
-	Message string
-	Code    int
-}
-
-func (e serverError) Log() {
-	log.Printf("\n\terror: %v \n\tmessage: %s \n\tcode: %d\n\n", e.Error, e.Message, e.Code)
-}
-
-type errHandler func(http.ResponseWriter, *http.Request) *serverError
-
-func (fn errHandler) ServeHTTP(w http.ResponseWriter, r *http.Request) {
-	if err := fn(w, r); err != nil {
-		err.Log()
-	}
-}
-
 func landingHandler(w http.ResponseWriter, r *http.Request) *serverError {
-	resp, err := http.Get(APIURL + "login")
+	resp, err := http.Get(frontendURI+apiPath+loginEndpoint)
 	if err != nil {
 		_ = landing.Render(w, login{})
 		return &serverError{
@@ -161,9 +91,9 @@ func landingHandler(w http.ResponseWriter, r *http.Request) *serverError {
 		}
 	}
 
-	au := &login{}
+	l := &login{}
 
-	err = json.Unmarshal(b, &au)
+	err = json.Unmarshal(b, &l)
 	if err != nil {
 		_ = landing.Render(w, login{})
 		return &serverError{
@@ -173,7 +103,7 @@ func landingHandler(w http.ResponseWriter, r *http.Request) *serverError {
 		}
 	}
 
-	_ = landing.Render(w, au)
+	_ = landing.Render(w, l)
 	return nil
 }
 
@@ -204,7 +134,7 @@ func resultHandler(w http.ResponseWriter, r *http.Request) *serverError {
 	v.Set("code", code)
 	v.Set("state", state)
 
-	resp, err := http.Get(APIURL + "playlist" + "?" + v.Encode())
+	resp, err := http.Get(frontendURI+apiPath+playlistEndpoint + "?" + v.Encode())
 	if err != nil {
 		_ = result.Render(w, playlist{})
 		return &serverError{
@@ -225,9 +155,9 @@ func resultHandler(w http.ResponseWriter, r *http.Request) *serverError {
 		}
 	}
 
-	list := &playlist{}
+	play := &playlist{}
 
-	err = json.Unmarshal(b, &list)
+	err = json.Unmarshal(b, &play)
 	if err != nil {
 		_ = result.Render(w, playlist{})
 		return &serverError{
@@ -237,7 +167,7 @@ func resultHandler(w http.ResponseWriter, r *http.Request) *serverError {
 		}
 	}
 
-	if blank.Is(string(list.URI)) {
+	if blank.Is(string(play.URI)) {
 		_ = result.Render(w, playlist{})
 		return &serverError{
 			Error:   err,
@@ -246,7 +176,7 @@ func resultHandler(w http.ResponseWriter, r *http.Request) *serverError {
 		}
 	}
 
-	_ = result.Render(w, list)
+	_ = result.Render(w, play)
 	return nil
 }
